Move dimensionless_3d_accessor to dimensionless folder
The file belongs with the other dimensionless components, not
in the eep subfolder. Update all include paths and BUILD deps.
