+++ src/option/BUILD.bazel
@@ -164,7 +164,7 @@ cc_library(
         "//src/option/table/bspline:bspline_3d_surface",
         "//src/option/table/chebyshev:chebyshev_3d_surface",
         "//src/option/table/dimensionless:dimensionless_builder",
-        "//src/option/table:dimensionless_3d_accessor",
+        "//src/option/table/dimensionless:dimensionless_3d_accessor",
         "//src/option/table:dimensionless_transform_3d",
         "//src/math:bspline_nd_separable",
         "//src/math/chebyshev:chebyshev_nodes",
