+++ src/option/table/dimensionless/BUILD.bazel
@@ -21,11 +21,23 @@ cc_library(
         "//src/option:american_option_batch",
         "//src/option:grid_spec_types",
         "//src/math:cubic_spline_solver",
-        "//src/option/table:dimensionless_3d_accessor",
+        ":dimensionless_3d_accessor",
         "//src/option/table:eep_decomposer",
         "//src/option/table:analytical_eep",
         "//src/math:bspline_nd_separable",
         "//src/option/table/bspline:bspline_surface",
+    ],
+    visibility = ["//visibility:public"],
+    strip_include_prefix = "/src/option/table/dimensionless",
+    include_prefix = "mango/option/table/dimensionless",
+)
+
+cc_library(
+    name = "dimensionless_3d_accessor",
+    hdrs = ["dimensionless_3d_accessor.hpp"],
+    deps = [
+        ":dimensionless_builder",
+        "//src/option/table:eep_decomposer",
     ],
     visibility = ["//visibility:public"],
     strip_include_prefix = "/src/option/table/dimensionless",
