+++ src/option/table/dimensionless/dimensionless_adaptive.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 #include "mango/option/table/dimensionless/dimensionless_builder.hpp"
 #include "mango/option/table/dimensionless/dimensionless_european.hpp"
-#include "mango/option/table/eep/dimensionless_3d_accessor.hpp"
+#include "mango/option/table/dimensionless/dimensionless_3d_accessor.hpp"
 #include "mango/option/table/eep/analytical_eep.hpp"
 #include "mango/math/bspline_nd_separable.hpp"
 #include "mango/option/table/bspline/bspline_surface.hpp"
